requirements.txt
google-generativeai
langchain-google-genai
chromadb
sqlalchemy
python-dotenv
pandas
sqlparse
mysql-connector-python 